+++ drizzle/0022_gorgeous_leopardon.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `sources` ADD `enabled` integer DEFAULT true NOT NULL;